+++ example/adapter-tests/collated.js
@@ -1,0 +1,6 @@
+module.exports = function collated(fractal) {
+    it('renders collated components collated', async () => {
+        const render = await fractal.components.find('@collated').render(null, null, { collate: true });
+        expect(render).toMatchSnapshot();
+    });
+};