+++ example/adapter-tests/index.js
@@ -1,0 +1,10 @@
+module.exports.components = require('./components');
+module.exports.tree = require('./tree');
+module.exports.renderTag = require('./render-tag');
+module.exports.docs = require('./docs');
+module.exports.render = require('./render');
+module.exports.notes = require('./notes');
+module.exports.include = require('./include');
+module.exports.context = require('./context');
+module.exports.collated = require('./collated');
+module.exports.path = require('./path');