+++ example/components/context/context.spec.js
@@ -1,0 +1,11 @@
+const { context } = require('../../adapter-tests');
+
+const fractal = require('../../fractal.config.js');
+
+describe('tree', () => {
+    beforeEach(async () => {
+        await fractal.load();
+    });
+
+    context(fractal);
+});