+++ example/components/include/include.spec.js
@@ -1,0 +1,11 @@
+const { include } = require('../../adapter-tests');
+
+const fractal = require('../../fractal.config.js');
+
+describe('include', () => {
+    beforeEach(async () => {
+        await fractal.load();
+    });
+
+    include(fractal);
+});