+++ example/components/notes/notes.spec.js
@@ -1,0 +1,11 @@
+const { notes } = require('../../adapter-tests');
+
+const fractal = require('../../fractal.config.js');
+
+describe('notes', () => {
+    beforeEach(async () => {
+        await fractal.load();
+    });
+
+    notes(fractal);
+});