+++ example/components/render-tag/render-tag.spec.js
@@ -1,0 +1,11 @@
+const { renderTag } = require('../../adapter-tests');
+
+const fractal = require('../../fractal.config.js');
+
+describe('render-tag', () => {
+    beforeEach(async () => {
+        await fractal.load();
+    });
+
+    renderTag(fractal);
+});