+++ example/components/render/render.spec.js
@@ -1,0 +1,11 @@
+const { render } = require('../../adapter-tests');
+
+const fractal = require('../../fractal.config.js');
+
+describe('render', () => {
+    beforeEach(async () => {
+        await fractal.load();
+    });
+
+    render(fractal);
+});